+++ src/graphics/view.rs
@@ -1,4 +1,4 @@
-use geom::{Vector, Rectangle, Transform};
+use crate::geom::{Vector, Rectangle, Transform};
 
 #[derive(Clone, Copy, Debug)]
 ///A view into the world, used as a camera and a viewport