+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Add a new ResizeStrategy: IntegerScale
 - Fix bug in `Rectangle::contains`
 - Fix transformed lines not displaying properly
+- Update to Rust 2018
 
 ## 0.3.2