+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## In-development
 
+## 0.3.3
+
 - Add a new ResizeStrategy: IntegerScale
 - Fix bug in `Rectangle::contains`
 - Fix transformed lines not displaying properly